Move GeoPointMapActivityTest to geo module
