+++ collect_app/src/main/AndroidManifest.xml
@@ -137,10 +137,11 @@ the specific language governing permissions and limitations under the License.
         <activity
             android:name="org.odk.collect.geo.GeoPointActivity"
             android:theme="@style/Theme.Collect"
-            tools:replace="android:theme"/>
+            tools:replace="android:theme" />
         <activity
             android:name="org.odk.collect.geo.GeoPointMapActivity"
-            android:theme="@style/Theme.Collect" />
+            android:theme="@style/Theme.Collect"
+            tools:replace="android:theme" />
         <activity android:name=".activities.GeoPolyActivity" />
         <activity android:name=".activities.FormMapActivity" />
         <activity android:name=".activities.BearingActivity" />
